GUI Improved and Log system started
Added functions to the GUI, using unicode to display symbols for pi and sqrt(). Changed Logger to accept those new inputs in the specific encoding. Created a file for Custom Error Handling types that we can define as need be, created IllegalArgumentError as a general Exception for bad argument passed. changed computePi() for the function to run fast enough that we can calculate the value of pi as needed and as required by the prof. Fixed Error handling in the GUI to avoid terminating the program for a simple syntax error.